+++ Documentation/GUI/Templates/GUI Component Template.lua
@@ -1,0 +1,70 @@
+-------------------------------------------------------------------------------
+-- Project      : ACP Studio
+-- Component    : <Component Name>
+-- Module       : <Module Name>
+-- File         : Core/GUI/<Component>/<Module>.lua
+--
+-- Purpose
+-- -------
+-- <Describe the purpose of the component.>
+--
+-- Responsibilities
+-- ----------------
+-- • <Responsibility 1>
+-- • <Responsibility 2>
+-- • <Responsibility 3>
+--
+-- Public API
+-- ----------
+-- Initialize(...)
+-- Shutdown()
+-- IsInitialized()
+-- <Additional Public Functions>
+--
+-- Certification
+-- -------------
+-- GUI-XXX : <Milestone Name>
+-------------------------------------------------------------------------------
+
+local Module = {}
+
+-------------------------------------------------------------------------------
+-- Private State
+-------------------------------------------------------------------------------
+
+-------------------------------------------------------------------------------
+-- Initializes the component.
+--
+-- Parameters
+-- ----------
+-- <Describe parameters, if any.>
+-------------------------------------------------------------------------------
+function Module.Initialize(...)
+
+end
+
+-------------------------------------------------------------------------------
+-- Shuts down the component.
+-------------------------------------------------------------------------------
+function Module.Shutdown()
+
+end
+
+-------------------------------------------------------------------------------
+-- Returns whether the component has been initialized.
+--
+-- Returns
+-- -------
+-- True if initialized, otherwise false.
+-------------------------------------------------------------------------------
+function Module.IsInitialized()
+
+end
+
+-------------------------------------------------------------------------------
+-- Public API
+-------------------------------------------------------------------------------
+
+-- Additional public functions are declared here.
+
+return Module